requirements.txt
flask==2.0.1
flask-cors==3.0.10
flask-talisman==0.8.1
flask-limiter==1.4
python-dotenv==0.19.0
gunicorn==20.1.0
plaid-python==7.2.1
requests==2.26.0
pandas==1.3.3